+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   potential path contained in equivalence labels.
 - `PartialSpecificationRuleExtractor` that extracts rules from partially built
   specifications and two subclasses with specific applications of that.
+- Make specification iterable
 
 ### Changed
 - If a rule in a specification cannot be sanity checked (e.g., counting is not